refactor(controllers)!: (#83) move logic to AuthController, delete UserController
